Put overworld decorators in their own package
