+++ src/main/java/net/glowstone/generator/decorators/overworld/CactusDecorator.java
@@ -1,6 +1,8 @@
-package net.glowstone.generator.decorators;
+package net.glowstone.generator.decorators.overworld;
 
 import java.util.Random;
+
+import net.glowstone.generator.decorators.BlockDecorator;
 
 import org.bukkit.Chunk;
 import org.bukkit.Material;
