+++ src/main/java/net/glowstone/generator/decorators/overworld/FlowerDecorator.java
@@ -1,10 +1,12 @@
-package net.glowstone.generator.decorators;
+package net.glowstone.generator.decorators.overworld;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
+
+import net.glowstone.generator.decorators.BlockDecorator;
 
 import org.bukkit.Chunk;
 import org.bukkit.Material;
